# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VMedTriage
  description: Controlled medical triage support pipeline with mandatory human approval
  version: 1.0.0
paths:
  /api/v1/register:
    post:
      summary: Register
      description: Create a patient account, or a nurse account with the private registration
        code.
      operationId: register_api_v1_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/login:
    post:
      summary: Login
      description: Authenticate an account and issue a short-lived JWT access token.
      operationId: login_api_v1_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/register:
    post:
      summary: Register Alias
      description: 'Alias của POST /register theo đặc tả Feature #3 (POST /auth/register).
        Giữ /register để tương thích ngược.'
      operationId: register_alias_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      summary: Login Alias
      description: 'Alias của POST /login theo đặc tả Feature #3 (POST /auth/login).
        Giữ /login để tương thích ngược.'
      operationId: login_alias_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/email-verification/confirm:
    post:
      summary: Confirm Email Verification
      operationId: confirm_email_verification_api_v1_auth_email_verification_confirm_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationConfirmRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/email-verification/resend:
    post:
      summary: Resend Email Verification
      operationId: resend_email_verification_api_v1_auth_email_verification_resend_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/password-reset/request:
    post:
      summary: Request Password Reset
      operationId: request_password_reset_api_v1_auth_password_reset_request_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/password-reset/confirm:
    post:
      summary: Confirm Password Reset
      operationId: confirm_password_reset_api_v1_auth_password_reset_confirm_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/change-password:
    post:
      summary: Change Password
      operationId: change_password_api_v1_auth_change_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/me:
    get:
      summary: Current User
      description: Return the user represented by the validated Bearer token.
      operationId: current_user_api_v1_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    put:
      summary: Update Current User
      operationId: update_current_user_api_v1_me_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/chat:
    post:
      summary: Chat
      description: 'Nhận tin nhắn tự do của bệnh nhân và chạy AGENT triệu chứng (`symptom_protocol/`).


        Trước đây endpoint này chạy pipeline rule-based `src/agents/graph.py`. Đã
        chuyển sang agent vì

        agent hỏi theo cụm/stage đúng tài liệu CS, chốt đỏ ngay khi phát hiện red
        flag, và mọi kết luận

        vẫn do rule engine THUẦN quyết định (LLM chỉ trích xuất field, không xếp mức
        khẩn cấp).


        Phiên mở ở LƯỢT MỞ, không ghim sẵn protocol nào: tin nhắn đầu là lời kể tự
        do, protocol

        (`fever` / `general`) chỉ được chọn sau khi đã trích xuất được lời kể đó.
        Ghim sẵn fever ở đây

        từng khiến người nhắn "tôi đau ngực, đi vài bước là hụt hơi" bị hỏi "bé hay
        người lớn, bao nhiêu

        tuổi" rồi đi hết bộ câu hỏi về sốt, và không luật đỏ nào quét được ca đó.


        HỢP ĐỒNG API KHÔNG ĐỔI (`ChatRequest`/`ChatResponse`) và case vẫn được ghi
        vào `case_store` qua

        `symptom_case_bridge` - hàng đợi điều dưỡng, lịch sử bệnh nhân và luồng duyệt
        HITL chạy y như cũ.

        `case_id` chính là `session_id` của phiên agent.'
      operationId: chat_api_v1_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/tools/{tool_name}/call:
    post:
      summary: Call Tool
      description: Call an external MCP tool through the registry.
      operationId: call_tool_api_v1_tools__tool_name__call_post
      parameters:
      - name: tool_name
        in: path
        required: true
        schema:
          type: string
          title: Tool Name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPToolCallRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPToolCallResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/cases/{case_id}:
    get:
      summary: Get Case
      operationId: get_case_api_v1_cases__case_id__get
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          title: Case Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriageCase'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cases/{case_id}/review:
    post:
      summary: Review Case
      operationId: review_case_api_v1_cases__case_id__review_post
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          title: Case Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NurseReviewRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NurseReviewResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    EmailVerificationConfirmRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        code:
          type: string
          maxLength: 6
          minLength: 6
          title: Code
      type: object
      required:
      - email
      - code
      title: EmailVerificationConfirmRequest
...
    Gender:
      type: string
      enum:
      - male
      - female
      - other
      - prefer_not_to_say
      title: Gender
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        email:
          type: string
          maxLength: 320
          minLength: 1
          title: Email
          description: Email, số điện thoại hoặc tên đăng nhập
        password:
          type: string
          maxLength: 128
          minLength: 1
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
      description: 'Đăng nhập bằng email, số điện thoại HOẶC tên đăng nhập.


        Giữ tên trường là `email` để không phá client hiện có, nhưng kiểu là `str`
        chứ không phải

        `EmailStr` - nếu ép EmailStr thì gõ tên đăng nhập sẽ bị chặn ngay ở tầng validate
        (422) và

        không bao giờ tới được bước tra cứu.'
...
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: MessageResponse
...
    RegisterRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        username:
          anyOf:
          - type: string
            maxLength: 50
            minLength: 3
          - type: 'null'
          title: Username
        phone_number:
          type: string
          maxLength: 32
          minLength: 9
          title: Phone Number
        full_name:
          type: string
          maxLength: 120
          minLength: 2
          title: Full Name
        date_of_birth:
          type: string
          format: date
          title: Date Of Birth
          default: '1900-01-01'
        gender:
          $ref: '#/components/schemas/Gender'
          default: prefer_not_to_say
        avatar_data_url:
          type: string
          maxLength: 2000000
          minLength: 30
          title: Avatar Data Url
          default: data:image/png;base64,AAAAAAAAAAAAAAAA
        password:
          type: string
          maxLength: 128
          minLength: 8
          title: Password
        confirm_password:
          type: string
          maxLength: 128
          minLength: 8
          title: Confirm Password
        terms_accepted:
          type: boolean
          title: Terms Accepted
        role:
          $ref: '#/components/schemas/UserRole'
          default: patient
        nurse_registration_code:
          anyOf:
          - type: string
            maxLength: 256
          - type: 'null'
          title: Nurse Registration Code
        professional_license:
          anyOf:
          - type: string
            maxLength: 160
          - type: 'null'
          title: Professional License
        workplace:
          anyOf:
          - type: string
            maxLength: 160
          - type: 'null'
          title: Workplace
        department:
          anyOf:
          - type: string
            maxLength: 160
          - type: 'null'
          title: Department
        bio:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Bio
      type: object
      required:
      - email
      - phone_number
      - full_name
      - password
      - confirm_password
      - terms_accepted
      title: RegisterRequest
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
        user:
          $ref: '#/components/schemas/UserResponse'
      type: object
      required:
      - access_token
      - expires_in
      - user
      title: TokenResponse
...
    UserResponse:
      properties:
        id:
          type: integer
          title: Id
        username:
          type: string
          title: Username
        email:
          type: string
          format: email
          title: Email
        full_name:
          type: string
          title: Full Name
        phone_number:
          type: string
          title: Phone Number
        date_of_birth:
          type: string
          format: date
          title: Date Of Birth
        gender:
          $ref: '#/components/schemas/Gender'
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        emergency_contact_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Emergency Contact Name
        emergency_contact_relationship:
          anyOf:
          - type: string
          - type: 'null'
          title: Emergency Contact Relationship
        emergency_contact_phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Emergency Contact Phone
        role:
          $ref: '#/components/schemas/UserRole'
        email_verified:
          type: boolean
          title: Email Verified
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - username
      - email
      - full_name
      - phone_number
      - date_of_birth
      - gender
      - address
      - emergency_contact_name
      - emergency_contact_relationship
      - emergency_contact_phone
      - role
      - email_verified
      - is_active
      - created_at
      title: UserResponse
...